Phase 4: Extract aircraft database and specs into aircraft-database.js module
Created aircraft-database.js (280 lines):
- Military aircraft database loading with caching (loadMilitaryDatabase)
- Military aircraft identification (isMilitaryAircraft, getMilitaryInfo)
- AIRCRAFT_TYPE_SPECS with 98 aircraft types (commercial, regional, business, GA, military, cargo)
- Aircraft specs lookup (getAircraftSpecs)

Updated app.js (11,180 → 10,951 lines):
- Added imports from aircraft-database.js
- Removed military database code (~85 lines)
- Removed aircraft specs database (~138 lines)
- Total reduction: 223 lines

Updated index.html:
- Added aircraft-database.js module loading
- Script order: config → constants → theme-manager → url-state-manager → aircraft-database → aircraft-svg-system → app

Updated 5-REFACTORING_LOG.md:
- Marked Phase 3 complete (commit 4b2a45a)
- Marked Phase 4 complete
- Documented extraction details
