+++ public/index.html
@@ -3928,6 +3928,9 @@
             // Load url-state-manager.js as a module (refactored URL state management)
             loadScript('url-state-manager.js', true);
 
+            // Load aircraft-database.js as a module (refactored aircraft database and specs)
+            loadScript('aircraft-database.js', true);
+
             // Load aircraft-svg-system.js
             loadScript('aircraft-svg-system.js');
